+++ src/lang/grammar.js
@@ -24,7 +24,7 @@ const {
     Parentheses,
     Pick,
     Pipeline,
-    Property,
+    ObjectEntry,
     Reference,
     SliceNotation,
     SortingFunction,
@@ -366,14 +366,14 @@ module.exports = {
         ],
         properties: createCommaList('properties', 'property'),
         property: [
-            ['ident', $$(Property($1, null))],
-            ['$', $$(Property(Current(), null))],  // do nothing, but collect stat (suggestions)
-            ['$ident', $$(Property(Reference($1), null))],
-            ['ident : e', $$(Property($1, $3))],
-            ['STRING : e', $$(Property(Literal($1), $3))],
-            ['NUMBER : e', $$(Property(Literal($1), $3))],
-            ['LITERAL : e', $$(Property(Literal($1), $3))],
-            ['[ e ] : e', $$(Property($2, $5))],
+            ['ident', $$(ObjectEntry($1, null))],
+            ['$', $$(ObjectEntry(Current(), null))],  // do nothing, but collect stat (suggestions)
+            ['$ident', $$(ObjectEntry(Reference($1), null))],
+            ['ident : e', $$(ObjectEntry($1, $3))],
+            ['STRING : e', $$(ObjectEntry(Literal($1), $3))],
+            ['NUMBER : e', $$(ObjectEntry(Literal($1), $3))],
+            ['LITERAL : e', $$(ObjectEntry(Literal($1), $3))],
+            ['[ e ] : e', $$(ObjectEntry($2, $5))],
             ['...', $$(Spread(null))],
             ['... query', $$(Spread($2))]
         ],